Namespaces ScrumProjects under scrum.
