+++ app/views/scrum/projects/_scrum_project.json.jbuilder
