+++ app/views/scrum/projects/index.json.jbuilder
