+++ app/views/scrum/projects/show.html.erb
