+++ app/views/scrum/projects/show.json.jbuilder
